+++ src/main/java/com/ssm/service/IUserService.java
@@ -9,8 +9,8 @@ import com.ssm.model.User;
  * @since  2015年9月28日
  */
 public interface IUserService {
-    User selectUserById(int userId) throws Exception;
-    User selectUserByIdNotAllData(int userId) throws Exception;
+    User selectUserById(long userId) throws Exception;
+    User selectUserByIdNotAllData(long userId) throws Exception;
     User selectUserByEmail(String email);
     User verifyUser(User user) throws Exception;
     boolean signUpUser(User user) throws Exception;
